refactor(openomni): split extension manager internals
